+++ CHANGELOG.md
@@ -1,3 +1,16 @@
+## [0.0.012] - 2022-02-08
+
+- Added kirki framework to handle wordpress cusomizer api. https://github.com/kirki-framework/kirki
+- Created Headless Horseman panel inside customizer to keep all theme related settings under one directory. 
+- Created an Editor color palette section to have possibility to add colors via UI. Once it is saved all color will appear in Gutenberg editor settings.
+
+### Added
+    - /inc/customizer/customizer.php
+    - /inc/customizer/settings.php
+    - /inc/customizer/sections/editor-color-pallete.php
+    - /inc/vendor/kirki
+
+
 ## [0.0.011] - 2022-01-14
 
 Added file to store functions that can be easily removed and are not affecting the theme directly.